refactor: add `testSolutions` function refactored from lines in `bin.ts`
