+++ src/index.ts
@@ -1,3 +1,3 @@
-export { testCppSolution } from "./test/cpp/index.js";
+export { testSolutions } from "./test/index.js";
 export { readYamlSchema, Schema } from "./test/schema.js";
 export { searchSolutions } from "./search.js";
